requirements.txt
reportlab
qrcode
Pillow
streamlit 